rename new event file to event-types
