+++ packages/back-end/src/events/notification-events.ts
@@ -1,6 +1,6 @@
 import { ApiExperiment, ApiFeature } from "../../types/openapi";
 import { NotificationEventPayload } from "./base-types";
-import { AuditableUserProperties } from "./events";
+import { AuditableUserProperties } from "./event-types";
 
 // region User
